extract Submission struct
